refactor: [WLEO-489] Accept ByteArray as issuerSignedContent instead of base64 string (#26)
## Short description
This PR changes the `DocRequested` data class which now accepts a
`ByteArray` instead of a base64 `String` for a `issuerSignedContent`
which is the credential CBOR.
Currently credentials are issued in `base64url` and a conversion to
base64 is needed which is then decoded by this library. By accepting a
`ByteArray` we can directly decode either base64url and base64 strings
in the bridge and send a `ByteArray` to this SDK.

## List of changes proposed in this pull request
- Change the `issuerSignedContent` from a base64 `String` to a
`ByteArray`;
- Remove the base64 `String` decoding in the `generateResponse` method;
- Update the example app by removing the encoding to base64;
- Update tests.

## How to test
Test a proximity presentation by running the example app contained in
this repository.